create new component for profile card in search result
