+++ services/backend/src/core/search/search.js
@@ -9,6 +9,7 @@ async function fuzzySearch(request, response) {
   const value = searchValue || "";
 
   const profiles = await utils.getAllProfiles(value, language, userId, request);
+
   const results = await utils.fuzzySearch(profiles, value);
 
   const responseData = utils.cleanResults(results);
