+++ services/frontend/src/components/resultsCard/ResultsCard.jsx
@@ -3,7 +3,7 @@ import { useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
 import { FormattedMessage } from "react-intl";
 import { Row } from "antd";
-import { map } from "lodash";
+import { map, property } from "lodash";
 import useAxios from "../../utils/useAxios";
 import ResultsCardView from "./ResultsCardView";
 import handleError from "../../functions/handleError";
@@ -43,7 +43,7 @@ const ResultsCard = () => {
       `api/profile/private/${id}?language=${locale}`
     );
 
-    setConnections(map(result.data.connections, "id"));
+    setConnections(map(result.data.connections, property("id")));
   }, [axios, id, locale]);
 
   useEffect(() => {
@@ -98,6 +98,7 @@ const ResultsCard = () => {
       connections={connections}
       addConnection={addConnection}
       removeConnection={removeConnection}
+      // getConnections={getConnections}
     />
   );
 };
